CNV-21086: Add 'Not migratable' badge VM Overview tab
Add 'Not migratable' badge to Status in the Details card
of the VM Overview tab, instead of 'Live migratable' field,
to clearly mark the VM that isn't live migratable.

This change follows the very recent design changes.

Fixes: https://issues.redhat.com/browse/CNV-21086
