+++ src/views/virtualmachines/details/tabs/details/components/grid/rightGrid/VirtualMachineDetailsRightGridLayout.tsx
@@ -82,18 +82,11 @@ const VirtualMachineDetailsRightGridLayout: React.FC<VirtualMachineDetailsRightG
     });
   };
 
-  const isMigratable = isLiveMigratable(vm, isSingleNodeCluster);
-
   return (
     <GridItem span={5}>
       <DescriptionList>
         <VirtualMachineDescriptionItem
-          descriptionData={
-            <VirtualMachineStatus
-              printableStatus={vm?.status?.printableStatus}
-              isMigratable={isMigratable}
-            />
-          }
+          descriptionData={<VirtualMachineStatus vm={vm} />}
           descriptionHeader={t('Status')}
           data-test-id={`${vm?.metadata?.name}-status`}
         />
